move assets inside the package
