+++ packages/mars-theme/assets/images/3bunnies-logo.gif
