+++ packages/mars-theme/assets/images/logo.png
