+++ packages/mars-theme/assets/images/rainbow-flag.png
